Get rid of a warning.
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@1081 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/sc2code/libs/uio/debug.c
@@ -397,7 +397,7 @@ debugCmdCat(DebugContext *debugContext, int argc, char *argv[]) {
 			totalWritten += numWritten;
 		} while (numInBuf > 0);
 	}
-	fprintf(debugContext->out, "[%d bytes]\n", totalWritten);
+	fprintf(debugContext->out, "[%u bytes]\n", (unsigned int) totalWritten);
 	
 	uio_close(handle);
 	return 0;
